+++ tests/test_plugin.py
@@ -41,9 +41,9 @@ def test_matches():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, "123", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "abc", (), 1, 0, 0, 0))
-    assert hu(Update(None, UpdateType.UPD)) == hr.SKIPPED
+    hu(Message(None, UpdateType.MSG, "123", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "abc", (), 1, 0, 0, 0, {}))
+    assert hu(Update(None, UpdateType.UPD, {})) == hr.SKIPPED
 
     assert debug.answers[1][0] == ("123", (), {})
     assert debug.answers[1][1] == ("abc", (), {})
@@ -61,9 +61,9 @@ def test_any_message():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, "123", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "abc", (), 1, 0, 0, 0))
-    assert hu(Update(None, UpdateType.UPD)) == hr.SKIPPED
+    hu(Message(None, UpdateType.MSG, "123", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "abc", (), 1, 0, 0, 0, {}))
+    assert hu(Update(None, UpdateType.UPD, {})) == hr.SKIPPED
 
     assert debug.answers[1][0] == ("123", (), {})
     assert debug.answers[1][1] == ("abc", (), {})
@@ -84,9 +84,9 @@ def test_any_unprocessed_message():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".abc", (), 1, 0, 0, 0))
-    assert hu(Update(None, UpdateType.UPD)) == hr.SKIPPED
+    hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".abc", (), 1, 0, 0, 0, {}))
+    assert hu(Update(None, UpdateType.UPD, {})) == hr.SKIPPED
 
     assert debug.answers[1][0] == (".123" * 2, (), {})
     assert debug.answers[1][1] == (".abc", (), {})
@@ -105,9 +105,9 @@ def test_any_update():
 
     app.add_plugin(pl)
 
-    hu(Update('a', UpdateType.UPD))
-    hu(Update('b', UpdateType.UPD))
-    assert hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0)) == hr.SKIPPED
+    hu(Update('a', UpdateType.UPD, {}))
+    hu(Update('b', UpdateType.UPD, {}))
+    assert hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0, {})) == hr.SKIPPED
 
     assert received[0] == 'a'
     assert received[1] == 'b'
@@ -132,9 +132,9 @@ def test_any_unprocessed_update():
 
     app.add_plugin(pl)
 
-    hu(Update('a', UpdateType.UPD))
-    hu(Update('b', UpdateType.UPD))
-    assert hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0)) == hr.SKIPPED
+    hu(Update('a', UpdateType.UPD, {}))
+    hu(Update('b', UpdateType.UPD, {}))
+    assert hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0, {})) == hr.SKIPPED
 
     assert received[0] == 'aa'
     assert received[1] == 'b'
@@ -163,9 +163,9 @@ def test_router_priority():
 
     app.add_plugin(pl)
 
-    hu(Update('a', UpdateType.UPD))
-    hu(Update('b', UpdateType.UPD))
-    assert hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0)) == hr.SKIPPED
+    hu(Update('a', UpdateType.UPD, {}))
+    hu(Update('b', UpdateType.UPD, {}))
+    assert hu(Message(None, UpdateType.MSG, ".123", (), 1, 0, 0, 0, {})) == hr.SKIPPED
 
     assert received[0] == 'aaa'
     assert received[1] == 'b'
@@ -182,17 +182,17 @@ def test_commands():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, ".echo 123", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".ec abc", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".echo", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".ec\n123", (), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, ".echo 123", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".ec abc", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".echo", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".ec\n123", (), 1, 0, 0, 0, {}))
 
-    hu(Message(None, UpdateType.MSG, ".ecabc", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".e cabc", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "abc", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "echo abc", (), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, ".ecabc", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".e cabc", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "abc", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "echo abc", (), 1, 0, 0, 0, {}))
 
-    assert hu(Update(None, UpdateType.UPD)) == hr.SKIPPED
+    assert hu(Update(None, UpdateType.UPD, {})) == hr.SKIPPED
 
     assert len(debug.answers[1]) == 4
     assert debug.answers[1][0] == (".echo 123", (), {})
@@ -211,9 +211,9 @@ def test_commands_with_spaces():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, " . echo 123", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ". echo 123", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".\nec abc", (), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, " . echo 123", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ". echo 123", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".\nec abc", (), 1, 0, 0, 0, {}))
 
     assert len(debug.answers[1]) == 3
     assert debug.answers[1][0] == (" . echo 123", (), {})
@@ -231,7 +231,7 @@ def test_command_full_body():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, ".echo abc\nabc\nabc", (), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, ".echo abc\nabc\nabc", (), 1, 0, 0, 0, {}))
 
     assert len(debug.answers[1]) == 1
     assert debug.answers[1][0] == ("abc\nabc\nabc", (), {})
@@ -252,13 +252,13 @@ def test_attachments():
     voice = Attachment("", "voice", "", None, "", None, "present")
     sticker = Attachment("", "sticker", "", None, "", None, "present")
 
-    hu(Message(None, UpdateType.MSG, "i", (image,), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "v", (voice,), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "vi", (voice, image), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "si", (sticker, image,), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, "s", (sticker,), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, "i", (image,), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "v", (voice,), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "vi", (voice, image), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "si", (sticker, image,), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, "s", (sticker,), 1, 0, 0, 0, {}))
 
-    assert hu(Update(None, UpdateType.UPD)) == hr.SKIPPED
+    assert hu(Update(None, UpdateType.UPD, {})) == hr.SKIPPED
 
     assert len(debug.answers[1]) == 4
     assert debug.answers[1][0] == ("i", (), {})
@@ -354,10 +354,10 @@ def test_sender_states():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0, {}))
 
     assert len(debug.answers[1]) == 3
     assert debug.answers[1].count(("ok:a", (), {})) == 2
@@ -391,10 +391,10 @@ def test_receiver_states():
 
     app.add_plugin(pl)
 
-    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0))
-    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0))
+    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".b", (), 1, 0, 0, 0, {}))
+    hu(Message(None, UpdateType.MSG, ".a", (), 1, 0, 0, 0, {}))
 
     assert len(debug.answers[1]) == 3
     assert debug.answers[1].count(("ok:a", (), {})) == 2